use qint64, we're talking nanoseconds here

+++ staging/kservice/tests/pluginlocator/plugintest.h
@@ -38,7 +38,7 @@ class PluginTest : public QCoreApplication
         void runMain();
         bool loadFromKService(const QString &name = "time");
         bool loadFromMetaData(const QString &pluginName = "plasma_engine_time");
-        void report(QList<int> timings, const QString &msg = "Test took ");
+        void report(QList<qint64> timings, const QString &msg = "Test took ");
 
     private:
         PluginTestPrivate* d;
